fix: move EtcdClientWrapper into factory package
